+++ Extensions/Announcers/GithubAnnouncements/GithubAnnouncements.csproj
@@ -51,16 +51,16 @@
     <Compile Include="BotIssuesExtensions.cs" />
     <Compile Include="BotPullRequestExtensions.cs" />
     <Compile Include="BotWatcherExtensions.cs" />
-    <Compile Include="CommitTask.cs" />
     <Compile Include="GithubAnnouncer.cs" />
-    <Compile Include="IGithubTask.cs" />
     <Compile Include="GitHub.cs" />
-    <Compile Include="ForksTask.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="PullRequestsTask.cs" />
     <Compile Include="SettingsExtensions.cs" />
     <Compile Include="StringExtensions.cs" />
+    <Compile Include="Tasks\CommitTask.cs" />
+    <Compile Include="Tasks\ForksTask.cs" />
+    <Compile Include="Tasks\IGithubTask.cs" />
     <Compile Include="Tasks\IssuesTask.cs" />
+    <Compile Include="Tasks\PullRequestsTask.cs" />
     <Compile Include="Tasks\WatchersTask.cs" />
   </ItemGroup>
   <ItemGroup>
