+++ Extensions/Announcers/GithubAnnouncements/Tasks/CommitTask.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.Composition;
 using System.Linq;
 using Jabbot;
 
-namespace GithubAnnouncements
+namespace GithubAnnouncements.Tasks
 {
     public class CommitTask : IGitHubTask
     {
